genai-engine/locust/requirements.txt
locust==2.40.2
pydantic==2.11.7
pandas==2.3.2
pyarrow==21.0.0
python-dotenv==1.1.1
